tui: split into multiple files for readability
