+++ tui/helpers.go
@@ -1,0 +1,11 @@
+package tui
+
+func clamp(n, min, max int) int {
+	if n < min {
+		return max
+	}
+	if n > max {
+		return min
+	}
+	return n
+}
